Add BuzzAdPopController
